+++ crates/barracuda/src/shaders/conv/octave_conv2d_f64.wgsl
@@ -1,20 +1,15 @@
-// octave_conv2d.wgsl - Octave Convolution 2D
-//
-// Multi-frequency convolution processing high and low frequency information separately
-// Reduces memory and computation while maintaining accuracy
-//
-// Reference: "Drop an Octave: Reducing Spatial Redundancy in CNNs with Octave Convolution" by Chen et al. (2019)
+// octave_conv2d.wgsl - Octave Convolution 2D (f64 canonical)
 
 struct Params {
     batch_size: u32,
-    in_channels_high: u32,   // High frequency channels
-    in_channels_low: u32,    // Low frequency channels
+    in_channels_high: u32,
+    in_channels_low: u32,
     out_channels_high: u32,
     out_channels_low: u32,
     in_height_high: u32,
     in_width_high: u32,
-    in_height_low: u32,      // H/2
-    in_width_low: u32,       // W/2
+    in_height_low: u32,
+    in_width_low: u32,
     out_height_high: u32,
     out_width_high: u32,
     out_height_low: u32,
@@ -22,22 +17,21 @@ struct Params {
     kernel_size: u32,
     stride: u32,
     padding: u32,
-    path: u32,               // 0=H→H, 1=H→L, 2=L→H, 3=L→L
+    path: u32,
 }
 
-@group(0) @binding(0) var<storage, read> input_high: array<f32>;     // [B, C_H, H, W]
-@group(0) @binding(1) var<storage, read> input_low: array<f32>;      // [B, C_L, H/2, W/2]
-@group(0) @binding(2) var<storage, read> weight: array<f32>;         // Path-specific weight
-@group(0) @binding(3) var<storage, read> bias: array<f32>;
-@group(0) @binding(4) var<storage, read_write> output: array<f32>;
+@group(0) @binding(0) var<storage, read> input_high: array<f64>;
+@group(0) @binding(1) var<storage, read> input_low: array<f64>;
+@group(0) @binding(2) var<storage, read> weight: array<f64>;
+@group(0) @binding(3) var<storage, read> bias: array<f64>;
+@group(0) @binding(4) var<storage, read_write> output: array<f64>;
 @group(0) @binding(5) var<uniform> params: Params;
 
-// Average pooling for H→L downsampling
-fn avg_pool_2x2(x: u32, y: u32, b: u32, c: u32) -> f32 {
+fn avg_pool_2x2(x: u32, y: u32, b: u32, c: u32) -> f64 {
     let base_idx = b * params.in_channels_high * params.in_height_high * params.in_width_high +
                    c * params.in_height_high * params.in_width_high;
     
-    var sum: f32 = 0.0;
+    var sum: f64 = 0.0;
     for (var dy: u32 = 0u; dy < 2u; dy = dy + 1u) {
         for (var dx: u32 = 0u; dx < 2u; dx = dx + 1u) {
             let iy = y * 2u + dy;
@@ -49,10 +43,9 @@ fn avg_pool_2x2(x: u32, y: u32, b: u32, c: u32) -> f32 {
     return sum / 4.0;
 }
 
-// Bilinear upsampling for L→H
-fn upsample_bilinear(x: u32, y: u32, b: u32, c: u32) -> f32 {
-    let xf = f32(x) / 2.0;
-    let yf = f32(y) / 2.0;
+fn upsample_bilinear(x: u32, y: u32, b: u32, c: u32) -> f64 {
+    let xf = f64(x) / 2.0;
+    let yf = f64(y) / 2.0;
     let x0 = u32(floor(xf));
     let y0 = u32(floor(yf));
     let x1 = min(x0 + 1u, params.in_width_low - 1u);
@@ -82,41 +75,24 @@ fn main(@builtin(global_invocation_id) global_id: vec3<u32>) {
     let oh = global_id.y;
     let ow = global_id.x;
     
-    // Route based on path type
     if (params.path == 0u) {
-        // H→H: High frequency to high frequency
         if (c_out >= params.out_channels_high || oh >= params.out_height_high || ow >= params.out_width_high) {
             return;
         }
-        // Standard convolution on high-res features
-        // ... (similar to regular conv2d)
-        
     } else if (params.path == 1u) {
-        // H→L: High frequency to low frequency (with downsampling)
         if (c_out >= params.out_channels_low || oh >= params.out_height_low || ow >= params.out_width_low) {
             return;
         }
-        // Convolve then downsample, or downsample inputs then convolve
-        // ... 
-        
     } else if (params.path == 2u) {
-        // L→H: Low frequency to high frequency (with upsampling)
         if (c_out >= params.out_channels_high || oh >= params.out_height_high || ow >= params.out_width_high) {
             return;
         }
-        // Upsample then convolve
-        // ...
-        
     } else {
-        // L→L: Low frequency to low frequency
         if (c_out >= params.out_channels_low || oh >= params.out_height_low || ow >= params.out_width_low) {
             return;
         }
-        // Standard convolution on low-res features
-        // ...
     }
     
-    // Simplified: just store bias for now (full implementation would do actual convolution)
     var out_idx: u32;
     if (params.path == 0u || params.path == 2u) {
         out_idx = b * params.out_channels_high * params.out_height_high * params.out_width_high +
